refactor: improve template package
