+++ packages/template/src/scenes/example.tsx
@@ -1,15 +1,23 @@
-import {Circle, makeScene2D} from '@motion-canvas/2d';
-import {createRef, waitFor, waitUntil} from '@motion-canvas/core';
+import {Rect, makeScene2D} from '@motion-canvas/2d';
+import {
+  all,
+  createRef,
+  easeInExpo,
+  easeInOutExpo,
+  waitFor,
+  waitUntil,
+} from '@motion-canvas/core';
 
 export default makeScene2D(function* (view) {
-  const circle = createRef<Circle>();
+  const rect = createRef<Rect>();
 
   view.add(
-    <Circle ref={circle} width={320} height={320} fill={'lightseagreen'} />,
+    <Rect ref={rect} size={320} radius={80} smoothCorners fill={'#f3303f'} />,
   );
 
-  yield* waitUntil('circle');
-  yield* circle().scale(2, 2).to(1, 2);
-
-  yield* waitFor(5);
+  yield* waitUntil('rect');
+  yield* rect().scale(2, 1, easeInOutExpo).to(1, 0.6, easeInExpo);
+  rect().fill('#ffa56d');
+  yield* all(rect().ripple(1));
+  yield* waitFor(0.3);
 });
